refactor(core): split permissioned_tool — extract auth/policy into unifi_core.permission
The @server.tool() decorator wiring stays in unifi-mcp-shared (MCP-coupled).
The auth/policy resolution logic moves to unifi_core.permission for reuse by apps/api
(the future rich HTTP API service).
